+++ README.md
@@ -32,7 +32,7 @@ console.log(JSON.stringify(ciEnvironment, null, 2))
 ### Java
 
 ```java
-import static io.cucumber.createmeta.DetectCiEnvironment.detectCiEnvironment;
+import static io.cucumber.cienvironment.DetectCiEnvironment.detectCiEnvironment;
 
 public class CiEnvironmentExample {
     public static void main(String[] args) {
