+++ java/src/main/groovy/io/cucumber/cienvironment/CiEnvironments.groovy
@@ -3,7 +3,7 @@ import groovy.json.JsonSlurper
 import java.nio.file.Files
 
 SimpleTemplateEngine engine = new SimpleTemplateEngine()
-def templateSource = new File(project.basedir, "src/main/groovy/io/cucumber/createmeta/CiEnvironments.gsp").getText()
+def templateSource = new File(project.basedir, "src/main/groovy/io/cucumber/cienvironment/CiEnvironments.gsp").getText()
 
 def jsonSlurper = new JsonSlurper()
 def ciEnvironments = jsonSlurper.parseText(new File(project.basedir, "../CiEnvironments.json").getText())
@@ -13,6 +13,6 @@ def toJava(s) {
 def binding = ["ciEnvironments": ciEnvironments, "toJava": this.&toJava]
 
 def template = engine.createTemplate(templateSource).make(binding)
-def file = new File(project.basedir, "target/generated-sources/ci-environments/java/io/cucumber/createmeta/CiEnvironments.java")
+def file = new File(project.basedir, "target/generated-sources/ci-environments/java/io/cucumber/cienvironment/CiEnvironments.java")
 Files.createDirectories(file.parentFile.toPath())
 file.write(template.toString(), "UTF-8")
