+++ java/src/main/groovy/io/cucumber/cienvironment/CiEnvironments.gsp
@@ -1,4 +1,4 @@
-package io.cucumber.createmeta;
+package io.cucumber.cienvironment;
 
 import static java.util.Arrays.asList;
