+++ java/src/main/java/io/cucumber/cienvironment/DetectCiEnvironment.java
@@ -1,4 +1,4 @@
-package io.cucumber.createmeta;
+package io.cucumber.cienvironment;
 
 import java.net.URI;
 import java.net.URISyntaxException;
