+++ java/src/main/java/io/cucumber/cienvironment/VariableExpression.java
@@ -1,4 +1,4 @@
-package io.cucumber.createmeta;
+package io.cucumber.cienvironment;
 
 import java.util.Map;
 import java.util.regex.Matcher;
