+++ javascript/tsconfig.json
@@ -8,7 +8,7 @@
       "@cucumber/gherkin": ["gherkin/javascript/src/index.ts"],
       "@cucumber/gherkin-streams": ["gherkin-streams/javascript/src/index.ts"],
       "@cucumber/gherkin-utils": ["gherkin-utils/javascript/src/index.ts"],
-      "@cucumber/create-meta": ["create-meta/javascript/src/index.ts"],
+      "@cucumber/ci-environment": ["ci-environment/javascript/src/index.ts"],
       "@cucumber/fake-cucumber": ["fake-cucumber/javascript/src/index.ts"],
       "@cucumber/query": ["query/javascript/src/index.ts"],
       "@cucumber/compatibility-kit": ["compatibility-kit/javascript/src/index.ts"],
